feat i1 gh workflow to just build exe
I've had trouble with using cargo run on the project itself. That
produces a null pointer exception in debian/ubuntu and some similar
but distinct problem when run on windows. I'm hoping github has
some setup that matches jdno's environment that results in
a valid binary. I'll grant that maybe cargo run is the culprit,
but I very hope not.