brk chg i0 revert to source of release 0.32.1
Compiling against August produces either a compile error because
the Geo library is incompatible or a null pointer exception when
actually run with the contemporary version. I don't trust that
jdno actually ran the binary after all the dependabot changes, just
ensured that it compiled.

So, this reverts to the version that I know he published. Hopefully,
these are all consistent.